requirements.txt
requests
python-dotenv
pandas
httpx
